fix: 画像レンダリング時の位置(offset)・回転・反転を適用
renderImage で transform の offsetX/offsetY が適用されておらず、
すべての画像が (0,0) に配置されていた問題を修正。
buildTransformAttr を共通モジュールに切り出し、image-renderer でも
shape-renderer と同等の transform 処理（translate, rotate, flip）を適用。
